Move train and test to model
